requirements.txt
streamlit
rdkit-pypi
opencv-python-headless
onnxruntime
scikit-learn
joblib
